feat: 전체 계좌 조회 요청 & 계좌 실명 조회 요청 (오픈뱅킹 연동 부분) 구현 및 테스트 (#39, SCRUM-131)

+++ ssok-account-service/src/main/java/kr/ssok/accountservice/entity/enums/AccountTypeCode.java
@@ -31,8 +31,8 @@ public enum AccountTypeCode {
         }
     }
 
-    public static AccountTypeCode fromIdx(Long idx) {
-        AccountTypeCode code = IDX_MAP.get(idx.intValue());
+    public static AccountTypeCode fromIdx(int idx) {
+        AccountTypeCode code = IDX_MAP.get(idx);
         if (code == null) {
             throw new IllegalArgumentException("Invalid AccountTypeCode idx: " + idx);
         }
